+++ src/events/GuildMemberAdd.ts
@@ -1,24 +1,22 @@
 import type { ArgsOf, Client } from "discordx";
 import { Discord, On } from "discordx";
+import { MembersModule } from "../modules/members/Members.module.js";
 import { joinSettings } from "../modules/members/joinNickname.js";
-import { logJoinLeaveEvents } from "../modules/members/logJoinLeaveEvents.js";
-import { updateMemberCount } from "../modules/members/updateMemberCount.js";
-import { upsertDbMember } from "../modules/members/upsertDbMember.js";
 
 @Discord()
 export class GuildMemberAdd {
-  @On()
+  @On({ event: "guildMemberAdd" })
   async guildMemberAdd([member]: ArgsOf<"guildMemberAdd">, client: Client) {
-    await logJoinLeaveEvents(member, "join");
+    MembersModule.logJoinLeaveEvents(member, "join");
 
     // create or update user with his roles
-    await upsertDbMember(member, "join");
+    await MembersModule.upsertDbMember(member, "join");
 
     // rules not yet accepted
     if (member.pending) return;
 
     // update user count channel
-    await updateMemberCount(member);
+    await MembersModule.updateMemberCount(member);
 
     await joinSettings(member);
   }
