+++ src/events/GuildMemberRemove.ts
@@ -1,7 +1,6 @@
 import type { ArgsOf, Client } from "discordx";
 import { Discord, On } from "discordx";
-import { updateMemberCount } from "../modules/members/updateMemberCount.js";
-import { upsertDbMember } from "../modules/members/upsertDbMember.js";
+import { MembersModule } from "../modules/members/Members.module.js";
 
 @Discord()
 export class GuildMemberRemove {
@@ -11,10 +10,10 @@ export class GuildMemberRemove {
     client: Client,
   ) {
     // create or update user with his roles
-    await upsertDbMember(member, "leave");
+    await MembersModule.upsertDbMember(member, "leave");
 
     // update user count channel
-    await updateMemberCount(member);
+    await MembersModule.updateMemberCount(member);
 
     //  await logJoinLeaveEvents(member, 'leave')
   }
